video-factory/requirements.txt
# FLUXION Video Factory — Dipendenze Python
# pip install -r requirements.txt --break-system-packages

# Google Cloud / Veo 3
google-cloud-aiplatform>=1.60.0
google-cloud-storage>=2.16.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0

# Video processing
ffmpeg-python>=0.2.0

# Text-to-speech
edge-tts>=6.1.9

# Upload
PyVimeo>=1.1.0
boto3>=1.34.0          # Cloudflare R2 (S3-compatible)

# HTTP / Utils
requests>=2.31.0
tqdm>=4.66.0

# QA opzionale (OCR brand check)
# pytesseract>=0.3.10
# Pillow>=10.0.0

# Runway fallback (opzionale)
# runwayml>=0.1.0

# Audio (opzionale, per analisi avanzata)
# pydub>=0.25.1
